test(recordings): keep wire tests within package budgets
